List: move indent and outdent helpers to a utils folder as plain functions
Move the indent and outdent logic and their shared helpers out of the
hooks folder into a utils folder, and turn the indent and outdent hooks
into plain functions that take the registry. Callers pass their own
registry, so the functions can run from anywhere. Indent's client ID is
optional and falls back to the first selected item.
