+++ packages/block-library/src/list-item/hooks/index.js
@@ -1,7 +1,4 @@
-export { default as useOutdentListItem } from './use-outdent-list-item';
-export { default as useIndentListItem } from './use-indent-list-item';
 export { default as useEnter } from './use-enter';
 export { default as useSpace } from './use-space';
 export { default as useMultiSelectTab } from './use-multi-select-tab';
 export { default as useMerge } from './use-merge';
-export { getIndentTarget, getOutdentTarget } from './indent-outdent-targets';
